Move name clash check into separate module
